refactor: move location and machine mapping to api lib
