+++ pkg/api/lib/v0/location.go
@@ -1,4 +1,4 @@
-package mapping
+package v0
 
 import (
 	"fmt"
