+++ pkg/api/lib/v0/node_test.go
@@ -6,15 +6,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-// TestValidLocation covers that a known threeport location is accepted and an
-// unknown one is rejected.
-func TestValidLocation(t *testing.T) {
-	// accepts a location present in the region map
-	assert.True(t, ValidLocation("Local"))
-	// rejects a location absent from the region map
-	assert.False(t, ValidLocation("NotARealLocation"))
-}
-
 // TestValidMachineSize covers that a known node size is accepted and an unknown
 // one is rejected.
 func TestValidMachineSize(t *testing.T) {
