+++ pkg/config/v0/aws_eks_kubernetes_runtime_instance.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/threeport/threeport/internal/kubernetes-runtime/mapping"
+	apilib "github.com/threeport/threeport/pkg/api/lib/v0"
 	api_v0 "github.com/threeport/threeport/pkg/api/v0"
 	client_lib "github.com/threeport/threeport/pkg/client/lib/v0"
 	client_v0 "github.com/threeport/threeport/pkg/client/v0"
@@ -162,7 +162,7 @@ func (a *AwsEksKubernetesRuntimeInstanceConfig) Create(
 	}
 
 	// get location for provider AWS region
-	location, err := mapping.GetLocationForAwsRegion(*awsEksKubernetesRuntimeInstanceValues.Region)
+	location, err := apilib.GetLocationForAwsRegion(*awsEksKubernetesRuntimeInstanceValues.Region)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get Threeport location for AWS region %s: %w", *awsEksKubernetesRuntimeInstanceValues.Region, err)
 	}
